requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
langchain==0.3.21
langchain-openai==0.3.9
langchain-community==0.3.20